Add files via upload
In this lab, I have worked with JavaScript, Bootstrap, HTML5, and jquery to manipulate the behavior of a form's input. It took care of Client-side validation where it checked if the required Information are correctly filled out or empty. It also checked the format of the input if both passwords match, the length of the email, and the password, the format of the password.