Add files via upload
In this Assignment, I used CURD Application for PHP MySQL, to create a server-based webpage to collect information from the customer. Here I created a form using HTML5 structure and added PHP to create, Update, read, and Delete data from a registration form for customers. I have used and created a database structure and uploaded it on the localhost server to set the table structure for the database. JSON Array was used to create the connection of the server port.